fix(ci): fold time_sleep into conc_registry to stay under disk budget
CI ran out of disk linking the workspace's test binaries — ld failed
on `session_budget` (a lex-store test, pre-existing) with
"No space left on device" before any test ran. ubuntu-latest gives
the workflow ~14 GB free after `free-disk-space@main` reclaims the
preinstalled stacks; my prior commit added two new test binaries in
lex-runtime (conc_registry + time_sleep), each of which re-links
arrow + polars (~250 MB stripped) — enough to tip the workspace over.

Merge the three `time.sleep` tests into `conc_registry.rs` so the
workspace adds one new test binary instead of two. Both files share
the same `run` helper shape; the consolidation is cosmetically clean
(time tests get their own `SLEEP_SRC` constant + a section header) and
file purpose remains primarily the conc-registry tests with the sleep
smokes as a small adjacent suite.

All 11 tests (8 conc + 3 sleep) pass in the merged binary.
