+++ crates/lex-runtime/tests/conc_registry.rs
@@ -245,3 +245,47 @@ fn registered_lists_names_sorted() {
         ].into()),
     );
 }
+
+// ── `time.sleep` smoke tests (#445) ─────────────────────────────────────
+//
+// Folded into this binary rather than a sibling test file so the
+// workspace adds one new test binary instead of two — each new binary
+// re-links lex-runtime + arrow + polars and the ubuntu-latest CI box
+// is already at its disk ceiling. Lower-bound assertions only on
+// elapsed time; CI hosts stall and we don't want a flaky upper bound.
+
+const SLEEP_SRC: &str = r#"
+import "std.time" as time
+import "std.datetime" as dt
+
+fn sleep_for(seconds :: Float) -> [time] Nil {
+  time.sleep(dt.duration_seconds(seconds))
+}
+"#;
+
+#[test]
+fn sleep_zero_duration_is_a_noop() {
+    let start = std::time::Instant::now();
+    let _ = run(SLEEP_SRC, "sleep_for", vec![Value::Float(0.0)]);
+    let elapsed = start.elapsed();
+    assert!(elapsed.as_millis() < 50,
+        "zero-duration sleep should return immediately, took {elapsed:?}");
+}
+
+#[test]
+fn sleep_50ms_elapses_real_wall_time() {
+    let start = std::time::Instant::now();
+    let _ = run(SLEEP_SRC, "sleep_for", vec![Value::Float(0.05)]);
+    let elapsed = start.elapsed();
+    assert!(elapsed.as_millis() >= 40,
+        "50ms sleep should take ~50ms, took {elapsed:?}");
+}
+
+#[test]
+fn sleep_negative_duration_is_a_noop() {
+    let start = std::time::Instant::now();
+    let _ = run(SLEEP_SRC, "sleep_for", vec![Value::Float(-1.0)]);
+    let elapsed = start.elapsed();
+    assert!(elapsed.as_millis() < 50,
+        "negative-duration sleep should return immediately, took {elapsed:?}");
+}
